requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
motor==3.3.2
python-dotenv==1.0.0
aiohttp==3.9.1
dnspython==2.4.2